edu/requirements.txt
numpy
pillow
matplotlib
